+++ src/edit/qt/tree/tag_tree_widget.cpp
@@ -262,4 +262,25 @@ namespace Invader::EditQt {
             return nullptr;
         }
     }
+    
+    std::optional<std::string> TagTreeWidget::get_selected_directory() const noexcept {
+        // If we have a tag selected, do nothing
+        if(this->get_selected_tag()) {
+            return std::nullopt;
+        }
+        
+        auto selected_items = this->selectedItems();
+        if(selected_items.size()) {
+            auto *item = selected_items[0];
+            std::string path;
+            while(item) {
+                path = (item->text(0) + INVADER_PREFERRED_PATH_SEPARATOR).toStdString() + path;
+                item = item->parent();
+            }
+            return path;
+        }
+        else {
+            return std::nullopt;
+        }
+    }
 }